pyproject.toml
[project]
name = "legora_dagster"
requires-python = ">=3.9,<3.14"
version = "0.1.0"
dependencies = ["dagster==1.11.7"]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

scrapy = [
    "beautifulsoup4==4.13.5",
    "boto3==1.40.20",
    "dateutils==0.6.12",
    "pymongo==4.14.1",
    "scrapy==2.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "legora_dagster"
registry_modules = [
    "legora_dagster.components.*",
]
